fix: 优化交互

+++ index.js
@@ -2,12 +2,12 @@
 
 (function(global) {
   const fs = require("fs");
-  const confirm = require("confirm-cli");
+  const inquirer = require("inquirer");
+  const chalk = require("chalk")
+  const logSymbols = require("log-symbols");
+
   const ArgumentParser = require("argparse").ArgumentParser;
   const unitsConfig = ["px", "px", "rem", "rem", "rpx", "rpx"];
-  console.log(
-    Array.from(new Set(queue(unitsConfig, 2).map(item => item.join(":"))))
-  );
   const parser = new ArgumentParser({
     version: "1.0.0",
     addHelp: true,
@@ -39,62 +39,40 @@
     console.log("\033[31mError: input参数不能为空！请运行-h查看帮助。\033[0m ");
   }
   function main(query) {
-    console.log(`正在读取 ${query.input}...`);
-    confirm(
-      "即将覆盖原文件，是否继续？",
-      function() {
+    if(!fs.existsSync(query.input)){
+      console.log(logSymbols.error, chalk.red(query.input + '目录不存在'));
+      return;
+    }
+    console.log(logSymbols.info, `正在读取 ${query.input}...`);
+    const promptList = [
+      {
+        type: 'confirm',
+        name: 'goOn',
+        message: "即将覆盖文件/文件夹" + chalk.yellow(query.input) + "，是否继续？"
+      }
+    ]
+    inquirer.prompt(promptList).then(answer => {
+      if(answer.goOn){
         DirHandler(query.input, function(path) {
           let inputStr = fs.readFileSync(path);
           let units = query.rule.split(":");
           let result = inputStr
             .toString()
             .replace(
-              new RegExp(`[ |:|\(]-?(\\d+(\\.\\d+)?)(${units[0]})`, "g"),
+              new RegExp(`[ |:|\(]-?((\\d+)?(\\.\\d+)?)(${units[0]})`, "g"),
               function(s, v, p, u) {
                 return s
-                  .replace(v, (v === "1" && query.x < 1) ? v : v * query.x)
+                  .replace(v, v * query.x)
                   .replace(u, units[1]);
               }
             );
           // yes
-          fs.writeFileSync(path, result);
+          if(result !== inputStr.toString()){
+            fs.writeFileSync(path, result);
+          }
         })
-        // DirHandler(query.input, function(path) {
-        //   let inputStr = fs.readFileSync(path);
-        //   let units = query.rule.split(":");
-        //   let result = inputStr
-        //     .toString()
-        //     .replace(
-        //       new RegExp(`[ |:|\(]-?(\\d+(\\.\\d+)?)(${units[0]})`, "g"),
-        //       function(s, v, p, u) {
-        //         return s
-        //           .replace(v, (v === "1" && query.x < 1) ? v : v * query.x)
-        //           .replace(u, units[1]);
-        //       }
-        //     );
-        //   // yes
-        //   fs.writeFileSync(path, result);
-        //   console.log(`已覆盖 ${path} 文件`);
-        //   let arr = path.split(".");
-        //   let fileName = arr.splice(0, arr.length - 1).join(".");
-        //   let type = arr[arr.length - 1];
-        //   let backFileName = type
-        //     ? fileName + "-" + Math.random() + "." + type
-        //     : fileName + "-" + Math.random();
-        //   fs.writeFileSync(
-        //     fileName + "-" + Math.random() + "." + type,
-        //     inputStr
-        //   );
-        //   console.log(`原文件已备份到 ${backFileName} 中`);
-        // });
-      },
-      function() {
-        
-      },
-      {
-        text: ["是", "否"]
       }
-    );
+    })
   }
   function DirHandler(path = "./", handler) {
     if (fs.lstatSync(path).isDirectory()) {
@@ -104,7 +82,7 @@
         DirHandler(folder + item, handler);
       });
     } else {
-      if(/\.(ux|css|less|scss|vue|js|jsx|ts|wxml|html)$/.test(path)) typeof handler === "function" && handler(path);
+      if(/\.(css|less|scss|sass|wxss|stylus|ux|vue|js|jsx|ts|tsx|wxml|html|php)$/.test(path)) typeof handler === "function" && handler(path);
     }
   }
   function queue(arr, size) {
